# portfolio/mcp-tool-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Agent Platform RAG API",
    "version": "0.3.0",
    "description": "OpenAPI contract for the Python Agent Platform knowledge-base RAG API."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local Agent Platform API"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        }
      }
    },
    "/documents": {
      "post": {
        "operationId": "ingestDocument",
        "summary": "Ingest a knowledge-base document",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentPayload"
              },
              "example": {
                "doc_id": "rag-guide",
                "title": "RAG Guide",
                "content": "Enterprise knowledge-base answers must include citations."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document ingested.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                },
                "example": {
                  "doc_id": "rag-guide",
                  "chunk_count": 1
                }
              }
            }
          }
        }
      }
    },
    "/ask": {
      "post": {
        "operationId": "askQuestion",
        "summary": "Ask a knowledge-base question with citations",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionPayload"
              },
              "example": {
                "question": "RAG 评估记录什么?"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent answer with citations or refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                },
                "example": {
                  "answer": "RAG 评估记录检索命中、引用覆盖、拒答和 trace。",
                  "refused": false,
                  "confidence": 0.9,
                  "citations": []
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "description": "Health status.",
            "example": "ok"
          }
        }
      },
      "DocumentPayload": {
        "type": "object",
        "required": ["doc_id", "title", "content"],
        "properties": {
          "doc_id": {
            "type": "string",
            "description": "Stable document id.",
            "example": "rag-guide"
          },
          "title": {
            "type": "string",
            "description": "Document title.",
            "example": "RAG Guide"
          },
          "content": {
            "type": "string",
            "description": "Plain-text document content.",
            "example": "Enterprise knowledge-base answers must include citations."
          }
        }
      },
      "IngestResponse": {
        "type": "object",
        "required": ["doc_id", "chunk_count"],
        "properties": {
          "doc_id": {
            "type": "string",
            "description": "Ingested document id.",
            "example": "rag-guide"
          },
          "chunk_count": {
            "type": "integer",
            "description": "Number of chunks created.",
            "example": 1
          }
        }
      },
      "QuestionPayload": {
        "type": "object",
        "required": ["question"],
        "properties": {
          "question": {
            "type": "string",
            "description": "User question.",
            "example": "RAG 评估记录什么?"
          },
          "session_id": {
            "type": "string",
            "description": "Optional multi-turn session id.",
            "example": "sess-001"
          }
        }
      },
      "AskResponse": {
        "type": "object",
        "required": ["answer", "refused", "confidence", "citations"],
        "properties": {
          "answer": {
            "type": "string",
            "description": "Final answer text.",
            "example": "RAG 评估记录检索命中、引用覆盖、拒答和 trace。"
          },
          "refused": {
            "type": "boolean",
            "description": "Whether the agent refused due to missing evidence.",
            "example": false
          },
          "confidence": {
            "type": "number",
            "description": "Confidence score from retrieval.",
            "example": 0.9
          },
          "citations": {
            "type": "array",
            "description": "Evidence citations backing the answer.",
            "example": []
          }
        }
      }
    }
  }
}